Give preference to CLI type when determining the default PHP exe
With PR #85 the environment factory started to check the
org.eclipse.php.debug.core.phpExe extension point for PHP executable if
one is not defined in the preferences.

There might be the case that there are PHP executables with various SAPI
types (CGI and CLI) registered in the extension point. Therefore, the
environment factory may pick up a CGI executable, which may cause
problems for the tools (like Composer) with "Maximum execution time of
30 seconds exceeded".

This patch modified the LaunchUtil.getPHPExecutable() method to give
preference to CLI executables.