ScriptLauncher should add LD_LIBRARY_PATH to the environment
This helps running correctly those PHP executable which pack the
required native libraries for the included PHP modules, so it is not
necessary to install them on the system.

For example, Zend Studio provides such built-in PHP executables and it
is having the LD_LIBRARY_PATH is important for running them correctly.

This patch uses the PHPLaunchUtilities.appendLibrarySearchPathEnv()
utility method from PDT to append the correct env var (LD_LIBRARY_PATH
for Linux, DYLD_FALLBACK_LIBRARY_PATH for Mac OS X) to the external
process.